Remove pointless variable
I was using $.each before and couldn't figure out how to return true
or return false, as in a jQuery foreach return is synonymous with
break and continue. I ditched the foreach and used a classical
for loop which solved my problem and is cleaner. I forgot to delete
the variable.